requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.6
h11==0.14.0
idna==3.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==1.4.45
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
pymysql==1.1.1
pydantic-settings==2.6.1
loguru==0.7.3
cffi==1.17.1
cryptography==38.0.4
pycparser==2.21
PyJWT==2.6.0
pytest-mock==3.14.0
httpx==0.28.1
